Add files via upload
Software Engineering class - Learning a new language in the given time and making a project given half of a semester to attempt to complete the game clicker game. IN CLASS WORK ONLY - no outside of class time was used. Minimal features in the game due to time restraints in the class